Moved "apollo-server-config.ts" from "server/misc" to "server"
