+++ src/server/backend-lambda.ts
@@ -1,5 +1,5 @@
 import { ApolloServer } from 'apollo-server-lambda';
-import { apolloServerConfig } from './misc/apollo-server-config';
+import { apolloServerConfig } from './apollo-server-config';
 
 const server = new ApolloServer(apolloServerConfig);
 export const handler = server.createHandler();
